backend/requirements.txt
fastapi[standard]
uvicorn[standard]
python-multipart
pdf2docx
python-docx
Pillow
moviepy
pydub
PyPDF2
